requirements.txt
streamlit
torch
torchvision
numpy
opencv-python
pillow
scikit-image
requests
pyyaml
imageio
imageio-ffmpeg